common: move merkle path helpers out of consensus
