+++ src/CMakeLists.txt
@@ -106,6 +106,7 @@ add_library(bitcoin_common STATIC EXCLUDE_FROM_ALL
   common/init.cpp
   common/interfaces.cpp
   common/license_info.cpp
+  common/merkle.cpp
   common/messages.cpp
   common/netif.cpp
   common/pcp.cpp
