+++ src/test/merkle_tests.cpp
@@ -2,6 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include <common/merkle.h>
 #include <consensus/merkle.h>
 #include <test/util/common.h>
 #include <test/util/random.h>
